#365 rename : DayType -> DateType로 파일 이름 변경
