+++ src/main/java/com/floney/floney/analyze/service/AssetServiceImpl.java
@@ -21,7 +21,7 @@ import java.util.Map;
 import static com.floney.floney.book.dto.constant.AssetType.BANK;
 import static com.floney.floney.book.dto.constant.AssetType.OUTCOME;
 import static com.floney.floney.book.dto.constant.CategoryEnum.FLOW;
-import static com.floney.floney.book.dto.constant.DayType.*;
+import static com.floney.floney.book.dto.constant.DateType.*;
 import static com.floney.floney.common.constant.Status.ACTIVE;
 
 @Service
@@ -78,7 +78,7 @@ public class AssetServiceImpl implements AssetService {
     @Transactional
     public void subtractAssetOf(final Long bookLineId) {
         final BookLine bookLine = bookLineRepository.findByIdWithCategories(bookLineId)
-                .orElseThrow(NotFoundBookLineException::new);
+            .orElseThrow(NotFoundBookLineException::new);
         if (!bookLine.includedInAsset()) {
             return;
         }
