+++ src/main/java/com/floney/floney/book/domain/vo/MonthLinesResponse.java
@@ -13,7 +13,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import static com.floney.floney.book.dto.constant.DayType.ONE_DAY;
+import static com.floney.floney.book.dto.constant.DateType.ONE_DAY;
 import static com.floney.floney.book.dto.process.MonthKey.toMonthKey;
 import static com.floney.floney.common.domain.vo.DateDuration.getStartAndEndOfMonth;
