+++ src/main/java/com/floney/floney/common/domain/vo/DateDuration.java
@@ -1,6 +1,6 @@
 package com.floney.floney.common.domain.vo;
 
-import com.floney.floney.book.dto.constant.DayType;
+import com.floney.floney.book.dto.constant.DateType;
 import com.floney.floney.book.dto.constant.ExcelDuration;
 import com.floney.floney.common.exception.book.LimitRequestException;
 import lombok.Builder;
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 import java.time.YearMonth;
 import java.time.format.DateTimeFormatter;
 
-import static com.floney.floney.book.dto.constant.DayType.*;
+import static com.floney.floney.book.dto.constant.DateType.*;
 
 @RequiredArgsConstructor
 @Getter
@@ -38,12 +38,12 @@ public class DateDuration {
 
     public static DateDuration getAssetDuration(LocalDate currentDate) {
         // 현재 날짜로부터 5개월 이전의 날짜 계산
-        LocalDate beforeMonth = getDateBeforeMonth(currentDate, FIVE_MOTH);
+        LocalDate beforeMonth = getDateBeforeMonth(currentDate, FIVE_MONTH);
         return new DateDuration(beforeMonth, currentDate);
     }
 
     // 현시점으로 부터 특정 개월 이후의 기간을 반환한다. ex. 1년 - 1월 1일 ~ 12월 31일
-    public static DateDuration getAfterMonthDuration(LocalDate firstDayOfMonth, DayType month) {
+    public static DateDuration getAfterMonthDuration(LocalDate firstDayOfMonth, DateType month) {
         LocalDate afterMonth = firstDayOfMonth.plusMonths(month.getValue());
         return new DateDuration(firstDayOfMonth, afterMonth.minusDays(1));
     }
@@ -73,7 +73,7 @@ public class DateDuration {
                 return getLastMonthDateDuration(targetDate);
             }
             case ONE_YEAR -> {
-                return getAfterMonthDuration(targetDate, DayType.ONE_YEAR_TO_MONTH);
+                return getAfterMonthDuration(targetDate, DateType.ONE_YEAR_TO_MONTH);
             }
             default -> {
                 throw new LimitRequestException();
@@ -89,7 +89,7 @@ public class DateDuration {
         return requestDate.withDayOfMonth(FIRST_DAY.getValue());
     }
 
-    private static LocalDate getDateBeforeMonth(LocalDate targetDate, DayType beforeMonth) {
+    private static LocalDate getDateBeforeMonth(LocalDate targetDate, DateType beforeMonth) {
         return targetDate.minusMonths(beforeMonth.getValue());
     }
